rename Comp
