+++ app/views/help/index.html.erb
@@ -9,4 +9,3 @@
     page: "Help"
   }) %>
 <% end %>
-
